fix(macos): harden LP64 runtime and loading

+++ src/aim_assist/aim_assist.h
@@ -189,7 +189,7 @@ void __cdecl AimAssist_DrawCenterBox(
     float clipHalfWidth,
     float clipHalfHeight,
     const float *color);
-void __cdecl AimAssist_DrawTargets(int64_t localClientNum, const float *color);
+void __cdecl AimAssist_DrawTargets(uint32_t localClientNum, const float *color);
 
 
 // aim_target_mp
